Updated Client Connector
