+++ Source/NexusForever.ClientConnector/Configuration/ClientConfiguration.cs
@@ -1,0 +1,7 @@
+﻿namespace NexusForever.ClientConnector
+{
+    public class ClientConfiguration
+    {
+        public string hostname { get; set; }
+    }
+}
