+++ Source/NexusForever.ClientConnector/Natives/PROCESS_INFORMATION.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace NexusForever.ClientConnector
+{
+    public struct PROCESS_INFORMATION
+    {
+        public IntPtr hProcess;
+        public IntPtr hThread;
+        public uint dwProcessId;
+        public uint dwThreadId;
+    }
+}
