+++ Source/NexusForever.ClientConnector/Natives/SECURITY_ATTRIBUTES.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace NexusForever.ClientConnector
+{
+    public struct SECURITY_ATTRIBUTES
+    {
+        public int length;
+        public IntPtr lpSecurityDescriptor;
+        public bool bInheritHandle;
+    }
+}
